+++ tests/conftest.py
@@ -235,7 +235,7 @@ class MockReactor:
 
     def monotonic(self):
         return self._monotonic
-
+    
     def pause(self, until):
         pass